Moves 007-openshift-yaml-health-probes to 020-quarkus-http-non-application-endpoints.
