--- 007-openshift-yaml-health-probes/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-quarkus.container-image.name=demo
