+++ 020-quarkus-http-non-application-endpoints/src/main/resources/application.properties
@@ -1,3 +1,4 @@
+quarkus.container-image.name=demo
 quarkus.application.name=non-application endpoints
 quarkus.http.root-path=/api
 quarkus.http.non-application-root-path=/q
